requirements.txt
Django~=2.0.7
django-filter~=2.2.0      
django-tables2~=2.1.0      